Number representation change

+++ src/main/java/com/weavechain/zk/bulletproofs/Allocated.java
@@ -3,11 +3,13 @@ package com.weavechain.zk.bulletproofs;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 @AllArgsConstructor
 public class Allocated {
 
     private final Variable variable;
 
-    private final Long assignment;
+    private final BigInteger assignment;
 }
